Split up corona tests into separate files
